Report added

+++ tsp.c
@@ -48,7 +48,7 @@ void draw_tsp(FILE * file_ptr, int n, float eps_width, float eps_height) {
 	for(i = 0; i < n; i++)
 	{
 		// calculate the radius
-		float radius = sqrt((eps_width * eps_height / (float) n) / 2 * PI) * SPACE_BETWEEN_CIRCLES_RATIO;
+		float radius = (float) (sqrt((eps_width * eps_height / (float) n) / 2 * PI) * SPACE_BETWEEN_CIRCLES_RATIO);
 		// calculate the correct x and y and shift for the edge positions
 		float x = (radius) + (cidades[i].x - min_x) / (max_x - min_x) * (eps_width - (2 * radius));
 		float y = (radius) + (cidades[i].y - min_y) / (max_y - min_y) * (eps_height - (2 * radius));
